Fetch a single model
When fetching one model instance by id, the result is an object. The line for (var i in data) would iterate on the object keys, which is not the desired behavior. Make a single element array out of the object and the rest of the code can stay the same to fetch one model or an entire collection.